repo/requirements.txt
Flask==3.0.0
cryptography==42.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
APScheduler==3.10.4
pytest==8.1.1
pytest-cov==5.0.0
requests==2.31.0
